Final Commit
Finished adding the New, Edit and Delete Board functionalities.
A couple issues are to be noted:
Whenever trying to edit and no name for the board has been entered, the api will return an error saying the name cannot be "undefined".
The biggest current issue is that whenever one logs in, the app can sometimes send you to a secondary page that states that the page should be reloaded. The app does not take in the credentials the first time the page is loaded, but everything seems to work fine after a reload.

Other than that, the app should work fine. The best way to debug the problems is to check Inspect Element!
Most of the errors I codded to appear in the log.

Hopefully everything works out fine.

As final words - I could say this was one of my hardest projects since I come from a very low to none knowledge background of React, and Lumen is quite the tricky framework to use (for me).